+++ src/pass/licm.rs
@@ -3,7 +3,9 @@
 //! Loop Independent Code Motion
 
 use crate::{
-    analysis::PredecessorTable, ir::prelude::*, opt::prelude::*, pass::gcse::DominatorTree,
+    analysis::{DominatorTree, PredecessorTable},
+    ir::prelude::*,
+    opt::prelude::*,
 };
 use std::collections::{HashMap, HashSet};