+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `UnitData` struct.
 - Add `Unit` and `UnitBuilder` structs.
+- Add `analysis` module.
 
 ### Changed
 - Fold `ControlFlowGraph`, `DataFlowGraph`, and `FunctionLayout` functions into `Unit` and `UnitBuilder`.
@@ -15,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `sig`, `dfg`, `cfg`, and `layout` fields of `UnitData` private.
 - Make `FunctionLayout` and `InstLayout` structs private.
 - Make `ControlFlowGraph` and `DataFlowGraph` structs private.
-- Factor out `TemporalRegionGraph` and `PredecessorTable` into new `analysis` module.
+- Factor out `TemporalRegionGraph`, `PredecessorTable`, and `DominatorTree` into `analysis` module.
 
 ### Removed
 - Remove `Function`, `Process`, and `Entity` structs.